Add files via upload
Q1: Write a PHP program to print your name and age to the browser. 
Q2: Write a PHP program which creates two variables that store your name and age and print them to the browser. 
Q3: Write a PHP program which creates constant variable that store your college name and print the value to the browser. 
Q4: Write a PHP program which creates the following variables and print them to the browser.
       $x=10;
       $y=7;